Holding update to Instructions to cover problem with max7219 lib
There is an incompatability with max7219 libuary used as part of the
Python code and the latest Rasbian release (29-11-2017). If you are
setting up from new use the previuos release and DO NOT run an apt-get
upgrade.

To fix this will require quite an extensive change to the code, I will
post an update when I have a stable version